Problem: Can't assign static addresses on GCP
Solution: Allow the user to specify 'createStaticExternal',
'createStaticInternal' to have the driver make new static addresses and
attach them when making an interface. Also allows for attaching a
specific static internal/external via  'attachInternal' and
'attachExternal' keys on an interface object. Where the value should be
the internal or external IPv4 address. At this time the driver doesn't
handle removal of static addresses, and it is assumed the user would
handle that via other means.