bugfix: check for `DeletionTimestamp` on ADD events
In case the MCM process stops during deletion, upon restart
the same terminating machine would be added to the normal
machine queue (as part of ADD event) and since its deletion
timestamp is set, it is not processed any further hence the
deletion never proceeds. By checking for a terminating machine
on ADD events, this is resolved.