Introduce machine termination queue to handle deletions (#964)
* re-use `machineCreateErrorHandler` in `CreationFlow`

Rather than explicitly setting the LastOperation value
and updating the status for multiple error codes, fall
back to using the already available ErrorHandler to
clean up the creation flow.

* feat: introducing machine termination queue

Isolated the machine deletion flow into a separate workerqueue
that only processes terminating machines, the remaining machine
reconciliation flow happens as is in the `machineQueue`. The
machine watch functions for `update` and `delete` events are
updated to account for the introduction of the new queue.

* bugfix: check for `DeletionTimestamp` on ADD events

In case the MCM process stops during deletion, upon restart
the same terminating machine would be added to the normal
machine queue (as part of ADD event) and since its deletion
timestamp is set, it is not processed any further hence the
deletion never proceeds. By checking for a terminating machine
on ADD events, this is resolved.

* nit: remove unnecessary guard around triggerDeletionFlow

* Address review comments

* machine.go: add tombstone handline for machine delete events
* machine.go: remove redundant check in creationflow
* machine_util.go: added docstring for createErrorHandler

* review: added check for successful casting

* nit: add explanatory comments and improve logged errors

* review: requeue on successful deletion as well