Adapt `MachineDeployment` controller
We skip any interaction with the target cluster by returning an error early if `InPlaceUpdate` is configured when running without a target cluster.
`InPlaceUpdate` only works with a target cluster – `Node` metadata needs to be updated for it.
Apart from that, there is no interaction with the target cluster in the `MachineDeployment` controller.