Adapt machine creation flow
During machine creation, we never set the `node` label on the `Machine` to disable all interaction with the target cluster.
With this, we skip waiting for the `Node` object to be registered.
We also don't create any bootstrap tokens.
Lastly, instead of tranisitioning to `Pending`, the `Machine` transitions to `Available` after creating the VM.