[GEP-28] Support managing machines without target cluster for `gardenadm bootstrap` (#1004)
* Add docs for running without target cluster

* Handle disabled target cluster in manager/controller/test setup

* Adapt `MachineDeployment` controller

We skip any interaction with the target cluster by returning an error early if `InPlaceUpdate` is configured when running without a target cluster.
`InPlaceUpdate` only works with a target cluster – `Node` metadata needs to be updated for it.
Apart from that, there is no interaction with the target cluster in the `MachineDeployment` controller.

* Adapt safety controller

If running without a target cluster, we can entirely disable the SafetyAPIServer and Node controllers.
Also, there is no need to annotate `Nodes` unmanaged by machine-controller-manager.

* Adapt machine creation flow

During machine creation, we never set the `node` label on the `Machine` to disable all interaction with the target cluster.
With this, we skip waiting for the `Node` object to be registered.
We also don't create any bootstrap tokens.
Lastly, instead of tranisitioning to `Pending`, the `Machine` transitions to `Available` after creating the VM.

* Adapt machine deletion flow

If running without a target cluster, we skip drain and volume attachment logic and go straight to VM deletion.
The `node` label does not exist, so we won't delete the `Node` object.

* Extract `constants.TargetKubeconfigDisabledValue`

* Move flag validation